Implement LoadsService for fetching load statistics; refactor LoadsTools to utilize service methods and improve error handling
